// src/app/interfaces/sports.ts
export interface ITennis {
    looser: string;
    numberOfSets: number;
    publicationDate: string;
    tournament: string;
    winner: string;
}

export interface IF1 {
    publicationDate: string;
    seconds: number;
    tournament: string;
    winner: string;
}

export interface INba {
    gameNumber: number;
    looser: string;
    mvp: string;
    publicationDate: string;
    tournament: string;
    winner: string; 
 }